+++ sdk/src/org.graalvm.launcher/src/org/graalvm/launcher/PolyglotLauncher.java
@@ -235,14 +235,13 @@ public final class PolyglotLauncher extends Launcher {
     }
 
     private void runScripts(List<Script> scripts, Context.Builder contextBuilder, String[] programArgs) {
-        Engine engine = Engine.create();
-        contextBuilder.engine(engine);
-        checkLanguage(mainLanguage, engine);
-        for (Script script : scripts) {
-            checkLanguage(script.languageId, engine);
-        }
         Script mainScript = scripts.get(scripts.size() - 1);
         try (Context context = contextBuilder.arguments(mainScript.getLanguage(), programArgs).build()) {
+            Engine engine = context.getEngine();
+            checkLanguage(mainLanguage, engine);
+            for (Script script : scripts) {
+                checkLanguage(script.languageId, engine);
+            }
             for (Script script : scripts) {
                 try {
                     Value result = context.eval(script.getSource());
@@ -252,6 +251,9 @@ public final class PolyglotLauncher extends Launcher {
                 } catch (PolyglotException e) {
                     if (e.isExit()) {
                         throw exit(e.getExitStatus());
+                    } else if (e.isGuestException()) {
+                        e.printStackTrace();
+                        throw exit(1);
                     } else {
                         throw abort(e);
                     }