+++ truffle/src/com.oracle.truffle.api.vm/src/com/oracle/truffle/api/vm/PolyglotImpl.java
@@ -56,6 +56,7 @@ import com.oracle.truffle.api.TruffleOptions;
 import com.oracle.truffle.api.impl.Accessor.EngineSupport;
 import com.oracle.truffle.api.impl.DispatchOutputStream;
 import com.oracle.truffle.api.impl.TruffleLocator;
+import com.oracle.truffle.api.interop.InteropException;
 import com.oracle.truffle.api.interop.Message;
 import com.oracle.truffle.api.interop.TruffleObject;
 import com.oracle.truffle.api.nodes.LanguageInfo;
@@ -168,6 +169,8 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
             return ((EngineException) e).e;
         } else if (e instanceof HostException) {
             return (HostException) e;
+        } else if (e instanceof InteropException) {
+            throw ((InteropException) e).raise();
         }
         return new HostException(e);
     }
@@ -263,14 +266,14 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
         public Env getEnvForLanguage(Object vmObject, String languageId, String mimeType) {
             PolyglotLanguageContext languageContext = (PolyglotLanguageContext) vmObject;
             PolyglotLanguageContext context = languageContext.context.findLanguageContext(languageId, mimeType, true);
-            context.ensureInitialized();
+            context.ensureInitialized(languageContext.language);
             return context.env;
         }
 
         @Override
         public Env getEnvForInstrument(Object vmObject, String languageId, String mimeType) {
             PolyglotLanguageContext context = PolyglotContextImpl.requireContext().findLanguageContext(languageId, mimeType, true);
-            context.ensureInitialized();
+            context.ensureInitialized(null);
             return context.env;
         }
 
@@ -333,7 +336,7 @@ public final class PolyglotImpl extends AbstractPolyglotImpl {
         public Env getEnvForInstrument(LanguageInfo info) {
             PolyglotLanguage language = (PolyglotLanguage) NODES.getEngineObject(info);
             PolyglotLanguageContext languageContext = PolyglotContextImpl.requireContext().contexts[language.index];
-            languageContext.ensureInitialized();
+            languageContext.ensureInitialized(null);
             return languageContext.env;
         }