+++ truffle/src/com.oracle.truffle.nfi.test/src/com/oracle/truffle/nfi/test/interop/BoxedPrimitive.java
@@ -41,4 +41,9 @@ public class BoxedPrimitive implements TruffleObject {
     public ForeignAccess getForeignAccess() {
         return BoxedPrimitiveMessageResolutionForeign.ACCESS;
     }
+
+    @Override
+    public String toString() {
+        return "boxed<" + primitive + ">";
+    }
 }